Update to 2.3.0

+++ docker/scripts/utils.sh
@@ -1,14 +1,23 @@
+#!/bin/bash
 SCRIPT_CONSTANTS=$HOME/scripts/constants.sh
 
-[ -f $SCRIPT_CONSTANTS ] && . $SCRIPT_CONSTANTS
-[ -f $SCRIPT_LOG ] && . $SCRIPT_LOG
+# shellcheck source=./scripts/constants.sh
+[ -f "$SCRIPT_CONSTANTS" ] && . "$SCRIPT_CONSTANTS"
+# shellcheck source=./scripts/log.sh
+[ -f "$SCRIPT_LOG" ] && . "$SCRIPT_LOG"
 
 function wait_file() {
   local file
+  local count
   file=$1
+  count=0
   while [[ ! -f "$file" ]]; do
     logd "waiting for $file..."
     sleep 5;
+    count=$((count + 1))
+    [[ $count -ge 50 ]] \
+      && loge "Waited too long, aborting..." && exit 1 \
+      || :
   done
 }
 
@@ -26,12 +35,12 @@ function exit_if_empty() {
 }
 
 function remove_sync_files() {
-  rm -rf $FOLDER_SYNC/*
+  rm -rf "${FOLDER_SYNC:?}/*"
 }
 
 function safety_check() {
   if [[ -f $FILE_SAFETY ]]; then
-    rm $FILE_SAFETY 
+    rm "$FILE_SAFETY "
   else
     loge "Safety enabled, run \`touch ${FOLDER_SYNC}/.new\`"
     exit 1
@@ -39,8 +48,9 @@ function safety_check() {
 }
 
 function capitalize_first_letter() {
-  ret=`tr '[:lower:]' '[:upper:]' <<< ${1:0:1}`${1:1}
-  echo $ret
+  local ret 
+  ret=$(tr '[:lower:]' '[:upper:]' <<< "${1:0:1}")${1:1}
+  echo "$ret"
 }
 
 function maybe_uninstall_app() {
@@ -48,8 +58,8 @@ function maybe_uninstall_app() {
   package_name=io.ptokens.p${NATIVE_SYMBOL}on${HOST_SYMBOL}
   case $TEE in
     strongbox )
-      $EXC_ADB uninstall $package_name 1>> /dev/null
-      [[ ! $? -eq 0 ]] \
+      
+      [[ ! $($EXC_ADB uninstall "$package_name" 1>> /dev/null) ]] \
         && logi "Failed to uninstall $package_name: maybe app doesn't exist..." \
         || logi "Uninstalling $package_name...done!"
       ;;
@@ -60,9 +70,7 @@ function maybe_install_app() {
   local apk_name=$HOME/p${NATIVE_SYMBOL}on${HOST_SYMBOL}-debug.apk
   case $TEE in
     strongbox )
-      $EXC_ADB install $apk_name >> /dev/null
-      
-      [[ ! $? -eq 0 ]] \
+      [[ ! $($EXC_ADB install "$apk_name" >> /dev/null) ]] \
         && loge "Failed to install $apk_name: abort!" && exit 1 \
         || logi "Installing $apk_name...done!"
 
@@ -75,8 +83,7 @@ function maybe_install_app() {
 function drop_logs() {
   case $TEE in
     vanilla )
-      rm -r $FOLDER_PROXY/logs 2> /dev/null
-      [[ ! $? -eq 0 ]] \
+      [[ ! $(rm -r "$FOLDER_PROXY/logs" 2> /dev/null) ]] \
         && logi "Failed to drop the logs, maybe folder doesn't exist..." \
         || logi "Dropping logs...done!"
       ;;
@@ -84,14 +91,24 @@ function drop_logs() {
 }
 
 function drop_sync_files() {
-  rm $FOLDER_SYNC/*.json 2> /dev/null
-  rm $FOLDER_SYNC/*.start 2> /dev/null
+  rm "$FOLDER_SYNC"/*.json 2> /dev/null
+  rm "$FOLDER_SYNC"/*.start 2> /dev/null
 
   [[ "$SKIP_SMART_CONTRACT_BYTECODE_GENERATION" -ne "1" ]] \
-    && rm -f $FOLDER_SYNC/smart-contract-bytecode \
+    && rm -f "$FOLDER_SYNC/smart-contract-bytecode" \
     || :
 
   logi "Removing sync files...done!"
+}
+
+function grant_permission() {
+  local package_name
+  local permission
+
+  package_name=$1
+  permission=$2
+
+  $EXC_ADB shell pm grant "$package_name" "$permission" >> /dev/null 
 }
 
 function maybe_set_write_storage_permissions() {
@@ -101,13 +118,13 @@ function maybe_set_write_storage_permissions() {
   permission=android.permission.WRITE_EXTERNAL_STORAGE
   case $TEE in
     strongbox )
-      $EXC_ADB shell pm grant $package_name $permission >> /dev/null 
-      [[ $? -eq 0 ]] \
-        && logi "WRITE_EXTERNAL_STORAGE permission set!" \
-        || loge "Failed to set the permission...aborting!"
+      if [[ $(grant_permission "$package_name" "$permission") ]]; then
+        logi "WRITE_EXTERNAL_STORAGE permission set!"
+      else
+        loge "Failed to set the permission...aborting!"
+      fi
       ;;
   esac
-    
 }
 
 function maybe_initialize_json_file() {
@@ -116,15 +133,16 @@ function maybe_initialize_json_file() {
   
   if [[ ! -f "$file" ]]; then
     logd "initialize $file"
-    echo "{}" > $file
+    echo "{}" > "$file"
   fi
 }
 
 function get_enclave_state() {
-  enclave_state=`cd $FOLDER_PROXY && $EXC_PROXY getEnclaveState`
+  enclave_state=$(cd "$FOLDER_PROXY" && $EXC_PROXY getEnclaveState)
   __enclave_state=$1
 
-  eval $__enclave_state="'$enclave_state'"
+  # shellcheck disable=SC2140
+  eval "$__enclave_state"="'$enclave_state'"
 }
 
 function touch_start_files() {
@@ -132,8 +150,8 @@ function touch_start_files() {
  
   symbol=$1
   
-  touch $FOLDER_SYNC/$symbol-broadcaster.start
-  touch $FOLDER_SYNC/$symbol-syncer.start
+  touch "$FOLDER_SYNC/$symbol-broadcaster.start"
+  touch "$FOLDER_SYNC/$symbol-syncer.start"
 }
 
 function check_safetyfile_exists() {
@@ -169,23 +187,21 @@ function add_bridge_info() {
   # TODO: this supports pbtc-on-eth only
 
   if [[ -f "$api_server_file" ]]; then
-    native_address=`cat $api_server_file \
-      | jq -r .ENCLAVE_PUBLIC_KEY`
-    smart_contract_address=`cat $api_server_file \
-      | jq -r .SMART_CONTRACT_ADDRESS`
+    # native_address=$(jq -r '.ENCLAVE_PUBLIC_KEY' "$api_server_file")
+    smart_contract_address=$(jq -r '.SMART_CONTRACT_ADDRESS' "$api_server_file")
 
-    bridge='''{
+    bridge='{
       "p'$NATIVE_SYMBOL'-on-'$HOST_SYMBOL'": {
         "NATIVE_TOKEN_ADDRESS": "'$public_key'",
         "HOST_TOKEN_ADDRESS": "'$smart_contract_address'",
         "NATIVE_SYMBOL": "'$NATIVE_SYMBOL'",
         "HOST_SYMBOL": "'$HOST_SYMBOL'"
       }
-    }'''
-    new_api_json=`cat $api_server_file \
-      | jq ".BRIDGES=$bridge"`
+    }'
+    
+    new_api_json=$(jq ".BRIDGES=$bridge" "$api_server_file")
 
     logd "$new_api_json"
-    echo "$new_api_json" > $api_server_file
+    echo "$new_api_json" > "$api_server_file"
   fi
 }
